- renamed optimizelyConfigHandler to ClientConfigHandler" - section name as constant in optimizelyFactory -
