+++ OptimizelySDK.Tests/App.config
@@ -7,7 +7,7 @@
 
   <configSections>
     <section name="optlySDKConfigSection"
-             type="OptimizelySDK.Config.OptimizelySDKConfigSection, OptimizelySDK, Version=3.2.0.0, Culture=neutral, PublicKeyToken=null" />
+             type="OptimizelySDK.OptimizelySDKConfigSection, OptimizelySDK, Version=3.2.0.0, Culture=neutral, PublicKeyToken=null" />
   </configSections>
   
   <optlySDKConfigSection>
