+++ OptimizelySDK/OptimizelySDK.csproj
@@ -71,7 +71,7 @@
     <Compile Include="AudienceConditions\OrCondition.cs" />
     <Compile Include="Bucketing\Bucketer.cs" />
     <Compile Include="Bucketing\Decision.cs" />
-    <Compile Include="Config\OptimizelyConfigHandler.cs" />
+    <Compile Include="ClientConfigHandler.cs" />
     <Compile Include="Entity\Attribute.cs" />
     <Compile Include="Entity\Audience.cs" />
     <Compile Include="AudienceConditions\BaseCondition.cs" />
